backend/pyproject.toml
[project]
name = "torch-chess-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-chess>=1.999",
    "torch>=2.5.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]
